# src/main/java/com/sbelan/walletapp/model/api/Transactions.java
package com.sbelan.walletapp.model.api;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Transactions {
    private List<Transaction> transactions;
}