# swagger/openapi.json
{
  "openapi":"3.0.1",
  "info":{
    "title":"OpenAPI definition",
    "version":"v0"
  },
  "servers":[{
    "url":"http://localhost:8080",
    "description":"Generated server url"
  }],
  "tags":[{
    "name":"wallet controller",
    "description":"wallet controller for maintaining user operations"
  }],
  "paths":{
    "/wallet/api/v1/debit":{
      "post":{
        "tags":["wallet controller"],
        "summary":"Debit operation",
        "operationId":"debit",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/DebitRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/DebitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/wallet/api/v1/credit":{
      "post":{
        "tags":["wallet controller"],
        "summary":"Credit operation",
        "operationId":"credit",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/CreditRequest"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/CreditResponse"
                }
              }
            }
          }
        }
      }
    },
    "/wallet/api/v1/history/{userId}":{
      "get":{
        "tags":["wallet controller"],
        "summary":"Get user transaction history",
        "operationId":"transactionHistory",
        "parameters":[{
            "name":"userId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
        }],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/Transactions"
                }
              }
            }
          }
        }
      }
    },
    "/wallet/api/v1/balance/{userId}":{
      "get":{
        "tags":["wallet controller"],
        "summary":"Get user balance",
        "operationId":"balance",
        "parameters":[{
            "name":"userId",
            "in":"path",
            "required":true,
            "schema":{
              "type":"integer",
              "format":"int64"
            }
        }],
        "responses":{
          "200":{
            "description":"OK",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/Balance"
                }
              }
            }
          }
        }
      }
    }
  },
  "components":{
    "schemas":{
      "DebitRequest":{
        "type":"object",
        "properties":{
          "userId":{
            "type":"integer",
            "format":"int64"
          },
          "sum":{
            "type":"number"
          }
        }
      },
      "DebitResponse":{
        "type":"object",
        "properties":{
          "responseMessage":{
            "type":"string"
          },
          "balance":{
            "type":"number"
          }
        }
      },
      "CreditRequest":{
        "type":"object",
        "properties":{
          "userId":{
            "type":"integer",
            "format":"int64"
          },
          "sum":{
            "type":"number"
          }
        }
      },
      "CreditResponse":{
        "type":"object",
        "properties":{
          "responseMessage":{
            "type":"string"
          },
          "balance":{
            "type":"number"
          }
        }
      },
      "Transaction":{
        "type":"object",
        "properties":{
          "id":{
            "type":"integer",
            "format":"int64"
          },
          "userId":{
            "type":"integer",
            "format":"int64"
          },
          "type":{
            "type":"string",
            "enum":["CREDIT","DEBIT"]
          },
          "sum":{
            "type":"number"
          },
          "date":{
            "type":"string",
            "format":"date-time"
          }
        }
      },
      "Transactions":{
        "type":"object",
        "properties":{
          "transactions":{
            "type":"array",
            "items":{
              "$ref":"#/components/schemas/Transaction"
            }
          }
        }
      },
      "Balance":{
        "type":"object",
        "properties":{
          "userId":{
            "type":"integer",
            "format":"int64"
          },
          "balance":{
            "type":"number"
          }
        }
      }
    }
  }
}